Derive vite define from references in env file + refactor yas-env
