+++ packages/yas-build-tools/package.json
@@ -5,6 +5,7 @@
   "license": "MIT",
   "devDependencies": {
     "@vitejs/plugin-react": "^4.0.1",
+    "@yas/env": "workspace:*",
     "rollup-plugin-visualizer": "^5.9.0",
     "vite": "^4.3.9",
     "vite-plugin-checker": "^0.5.6"
