+++ packages/yas-build-tools/vite.js
@@ -3,6 +3,7 @@ const react = require("@vitejs/plugin-react");
 const { visualizer } = require("rollup-plugin-visualizer");
 const { defineConfig } = require("vite");
 const { default: checker } = require("vite-plugin-checker");
+const { getEnvFile, getReferencesInEnvFile } = require("@yas/env/utils");
 
 // NOTE This is a JS file because it was much simpler than configuring
 // Vite & node to load non-prebuilt ESM files via the workspace:* monorepo directive.
@@ -14,22 +15,19 @@ function createYasViteConfig({ analyze } = {}) {
       checker({ typescript: true }),
       determineVisualizerPlugin(analyze),
     ].filter(Boolean),
-    define: exposeEnvToDefine(),
+    define: createDefineRepresentingEnvFile(),
   });
 }
 
 // We use define to normalize and use process.env even in vite apps.
 // This is to simplify tooling that need to import from our env file convention (see yas-env).
+// We automatically expose all env variables that are referenced in the env file.
 
-// Note on safety: Since we enforce an env var file convention, we are not as worried about
-// accidentally exposing secrets via process.env, since it's really easy to overview what env vars are used.
-// And one has to explicitly reference an env var by name to expose it. So just don't expose secrets.
-
-function exposeEnvToDefine() {
+function createDefineRepresentingEnvFile() {
   const { env } = process;
   const define = {};
-  for (const key in env) {
-    define[`process.env.${key}`] = JSON.stringify(env[key]);
+  for (const envVarName of getReferencesInEnvFile(getEnvFile(process.cwd()))) {
+    define[`process.env.${envVarName}`] = JSON.stringify(env[envVarName]);
   }
   return define;
 }
