+++ packages/yas-env/tsconfig.json
@@ -1,1 +1,6 @@
-{ "extends": "@yas/tsconfig/node.json" }
+{
+  "extends": "@yas/tsconfig/node.json",
+  "compilerOptions": {
+    "allowJs": true
+  }
+}
